[refactor] player=> CheckersPlayer
