+++ src/engine/CMakeLists.txt
@@ -1,6 +1,6 @@
 # use this to find in main "game" dir
 add_library(CheckersEngine
     CheckersEngine.cpp
-    player.cpp
+    CheckersPlayer.cpp
 )
