+++ src/include/CheckersPlayer.h
